Version 1.1 - Improved Sequence Curation under Multi-Sequence Input operation
Previous versions had multi-sequence input tacked on as a cool bonus feature, but it was not very well thought out or tested thoroughly. I ran a few iterations of AP-LASR with multi-sequence input to fix bugs in how the final dataset was curated. Specifically, hamming distance was not well defined under this mode and the post-alignment processing was too aggressive. Now hamming distance is defined and, under multi-sequence input, the number of iterations over the alignment during post-alignment processing is tunable by the user.

I made a few updates to the documentation as well.